testing/web-platform/tests/progress-events/tests/submissions/Samsung/security-consideration.sub.html

<!doctype html>
<html>
  <head>
    <title>ProgressEvent: security consideration</title>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <link rel="help" href="http://dvcs.w3.org/hg/progress/raw-file/tip/Overview.html#security-considerations" data-tested-assertations="/following-sibling::p" />
  </head>
  <body>
    <div id="log"></div>
    <script>
      var test = async_test();

      test.step(function() {
        var xhr = new XMLHttpRequest();

        xhr.onprogress = function(pe) {
          test.step(function() {
            if(pe.type == "progress") {
              assert_unreached("MUST NOT dispatch progress event.");
            }
          }, "Progress event named progress is not dispatched.");
        }
        xhr.onload = function(pe) {
          test.step(function() {
            if(pe.type == "load") {
              assert_unreached("MUST NOT dispatch progress event.");
            }
          }, "Progress event named load is not dispatched.");
        }
        xhr.onerror = function(pe) {
          test.step(function() {
            if(pe.type == "error") {
              assert_equals(pe.loaded, 0, "loaded is zero.");
              assert_false(pe.lengthComputable, "lengthComputable is false.");
              assert_equals(pe.total, 0, "total is zero.");
            }
          }, "Progress event named error does not reveal information.");
        }
        xhr.onloadend = function(pe) {
          test.done();
        }
        xhr.open("GET", "http://{{host}}:{{ports[http][1]}}/progress-events/tests/submissions/Samsung/resources/img.jpg", true);
        xhr.send(null);
      })
    </script>
  </body>
</html>
